login somente com o token

+++ app/Http/Controllers/Api/UsersController.php
@@ -10,6 +10,7 @@ use App\Models\LinkStats;
 use App\Models\User;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
+use Laravel\Sanctum\PersonalAccessToken;
 
 class UsersController extends Controller
 {
@@ -35,16 +36,50 @@ class UsersController extends Controller
 
     public function login(Request $request)
     {
-        $credentials = $request->only(['email', 'password']);
-        if (!Auth::attempt($credentials)) {
-            return response()->json('Incorrect email and / or password', 401);
+        if ($request->token) {
+            $personalAccessToken = PersonalAccessToken::findToken($request->token);
+
+            // Se encontrar o token
+            if ($personalAccessToken) {
+                // Obtém o usuário associado ao token
+                $user = $personalAccessToken->tokenable;
+
+                // Se encontrar o usuário
+                if ($user) {
+                    // Faz logout do usuário (caso já esteja autenticado)
+                    Auth::guard('web')->logout();
+
+                    // Autentica o usuário
+                    Auth::guard('web')->login($user);
+
+                    // Gera um novo token de acesso pessoal para o usuário
+                    $token = $user->createToken('token')->plainTextToken;
+
+                    // Retorna o token
+                    return response()->json($token, 200);
+                }
+            }
+        } else {
+            $credentials = $request->only(['email', 'password']);
+            if (!Auth::attempt($credentials)) {
+                return response()->json('Incorrect email and / or password', 401);
+            }
+
+            /** @var \App\Models\User $user **/
+            $user = Auth::user();
+            $user->tokens()->delete();
+            $token = $user->createToken('token');
+
+            return response()->json($token->plainTextToken, 200);
         }
+    }
 
-        /** @var \App\Models\User $user **/
-        $user = Auth::user();
-        $user->tokens()->delete();
-        $token = $user->createToken('token');
+    public function logout(Request $request)
+    {
+        $request->user()->tokens()->delete();
 
-        return response()->json($token->plainTextToken);
+        return response()->json([
+            'message' => 'Successfully logged out'
+        ]);
     }
 }
